+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allow JSONata expressions as input
+
 
 ## [0.1.1] - 2020-11-06 ![Relative date](https://img.shields.io/date/1604690140?label=)
 ### Added